feat(Dockerfile): add inkscape
I could not compile my Beamer presentation that used SVG images, as the
localleaf stack derives a bit from the Overleaf stack.
As it turns out (from the LaTeX log messages), the missing depency was
inkscape, to manage the images.

This commit adds inkscape, so that LaTeX libraries that manipulate
images now compile.